% Auto-generated: do not edit by hand
\name{dashSplitPane}

\alias{dashSplitPane}

\title{DashSplitPane component}

\description{

}

\usage{
dashSplitPane(children=NULL, id=NULL, allowResize=NULL, className=NULL, primary=NULL,
minSize=NULL, maxSize=NULL, defaultSize=NULL, size=NULL, split=NULL, style=NULL,
resizerStyle=NULL, paneClassName=NULL, pane1ClassName=NULL, pane2ClassName=NULL,
paneStyle=NULL, pane1Style=NULL, pane2Style=NULL, resizerClassName=NULL,
step=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{allowResize}{Logical. }

\item{className}{Character. }

\item{primary}{A value equal to: 'first', 'second'. }

\item{minSize}{Character | numeric. }

\item{maxSize}{Character | numeric. }

\item{defaultSize}{Character | numeric. }

\item{size}{Character | numeric. }

\item{split}{A value equal to: 'vertical', 'horizontal'. }

\item{style}{}

\item{resizerStyle}{}

\item{paneClassName}{Character. }

\item{pane1ClassName}{Character. }

\item{pane2ClassName}{Character. }

\item{paneStyle}{}

\item{pane1Style}{}

\item{pane2Style}{}

\item{resizerClassName}{Character. }

\item{step}{Numeric. }
}
